eliminated SkillLevelPanel from class hierarchy and replaced that class with AvailablePanel for following reasons:
AvailablePanel already existed but only in Visual Studio's  solution explorer as a filter that housed skilllevelpanel and other classes

the initial idea behind skilllevelpanel existing as its own class/panel no longer applies to how data flows in program. The name implies that somehow data belonging to skill level classes derived from unit.h only flowed within it, but again, that's no longer the case as additional functionality has caused skill level data to be sent to other places in program structure.

Instead of forcing the new ways in which data could be sorted to adhere to skilllevelpanel's exclusive hold on skilllevel data flow, it makes more sense to abandon the idea and allow availablepanel to take over. The name is easier to intuit its function from as well; all classes in availablepanel have to do with providing the user with the "units" and abilities that are available given the selected criteria.